+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/BankAccountRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.BankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface BankAccountRepository extends JpaRepository<BankAccount, String> {
 
     @Query("SELECT COALESCE(SUM(bankAccountLedgerEntry.amount), 0) FROM BankAccountLedgerEntry bankAccountLedgerEntry WHERE bankAccountLedgerEntry.bankAccount = :bankAccount")
-    public double balance(@Param("bankAccount") BankAccount bankAccount);
+    public BigDecimal balance(@Param("bankAccount") BankAccount bankAccount);
 }
