+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/CustomerRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface CustomerRepository extends JpaRepository<Customer, String> {
 
     @Query("SELECT COALESCE(SUM(customerLedgerEntry.amount), 0) FROM CustomerLedgerEntry customerLedgerEntry WHERE customerLedgerEntry.customer = :customer")
-    public double balance(@Param("customer") Customer customer);
+    public BigDecimal balance(@Param("customer") Customer customer);
 }
