+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/ItemRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface ItemRepository extends JpaRepository<Item, String> {
 
     @Query("SELECT COALESCE(SUM(itemLedgerEntry.quantity), 0) FROM ItemLedgerEntry itemLedgerEntry WHERE itemLedgerEntry.item = :item")
-    public double inventory(@Param("item") Item item);
+    public BigDecimal inventory(@Param("item") Item item);
 }
