+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/PostedSalesCreditMemoRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.PostedSalesCreditMemo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,6 +12,6 @@ import org.springframework.stereotype.Repository;
 public interface PostedSalesCreditMemoRepository extends JpaRepository<PostedSalesCreditMemo, String> {
 
     @Query("SELECT COALESCE(SUM(postedSalesCreditMemoLine.amount), 0) FROM PostedSalesCreditMemoLine postedSalesCreditMemoLine WHERE postedSalesCreditMemoLine.postedSalesCreditMemoLineId.postedSalesCreditMemo = :postedSalesCreditMemo")
-    public double amount(@Param("postedSalesCreditMemo") PostedSalesCreditMemo postedSalesCreditMemo);
+    public BigDecimal amount(@Param("postedSalesCreditMemo") PostedSalesCreditMemo postedSalesCreditMemo);
 
 }
