+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/PostedSalesOrderRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.PostedSalesOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,6 +12,6 @@ import org.springframework.stereotype.Repository;
 public interface PostedSalesOrderRepository extends JpaRepository<PostedSalesOrder, String> {
 
     @Query("SELECT COALESCE(SUM(postedSalesOrderLine.amount), 0) FROM PostedSalesOrderLine postedSalesOrderLine WHERE postedSalesOrderLine.postedSalesOrderLineId.postedSalesOrder = :postedSalesOrder")
-    public double amount(@Param("postedSalesOrder") PostedSalesOrder postedSalesOrder);
+    public BigDecimal amount(@Param("postedSalesOrder") PostedSalesOrder postedSalesOrder);
 
 }
