+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/VendorRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.Vendor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface VendorRepository extends JpaRepository<Vendor, String> {
 
     @Query("SELECT COALESCE(SUM(vendorLedgerEntry.amount), 0) FROM VendorLedgerEntry vendorLedgerEntry WHERE vendorLedgerEntry.vendor = :vendor")
-    public double balance(@Param("vendor") Vendor vendor);
+    public BigDecimal balance(@Param("vendor") Vendor vendor);
 }
